blog/requirements.txt
asgiref==3.4.1
boto3==1.20.23
botocore==1.23.23
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
jmespath==0.10.0
mysqlclient==2.0.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
